+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # ChangeLog
 
-* add xei lookup button.
-* fix multiblock formed issue
+* disable copy of non multiple traits
+* add recipe logic settings
+* fix create multiblock generator infinite rotation.